Some typing tidy up, removing alternatives as these are likely to cause confusion.
