+++ packages/@n8n/ai-workflow-builder.ee/src/prompts/shared/node-recommendations/audio-generation.ts
@@ -1,5 +1,7 @@
-import type { NodeRecommendationDocument } from '../../../types/node-recommendations';
-import { RecommendationCategory } from '../../../types/recommendation-category';
+import {
+	type NodeRecommendationDocument,
+	RecommendationCategory,
+} from '@/types/node-recommendations';
 
 export const audioGenerationRecommendation: NodeRecommendationDocument = {
 	category: RecommendationCategory.AUDIO_GENERATION,
