+++ packages/@n8n/ai-workflow-builder.ee/src/types/index.ts
@@ -11,10 +11,7 @@ export type * from './utils';
 export type * from './categorization';
 export type * from './best-practices';
 export type * from './node-guidance';
-export type * from './node-recommendations';
 
-// Re-export recommendation categories (includes both types and runtime values)
-export * from './recommendation-category';
-
-// Re-export web/templates (includes both types and runtime values)
+// exports with enumerations, export types and enums
 export * from './web/templates';
+export * from './node-recommendations';
